Studio photo added

+++ src/main/java/by/bsuir/tattoo4u/controller/PostController.java
@@ -2,13 +2,8 @@ package by.bsuir.tattoo4u.controller;
 
 import by.bsuir.tattoo4u.dto.request.PostRequestDto;
 import by.bsuir.tattoo4u.dto.response.PostResponseDto;
-import by.bsuir.tattoo4u.entity.PhotoUpload;
-import by.bsuir.tattoo4u.entity.Post;
-import by.bsuir.tattoo4u.entity.User;
-import by.bsuir.tattoo4u.service.PostService;
-import by.bsuir.tattoo4u.service.ServiceException;
-import by.bsuir.tattoo4u.service.TokenService;
-import by.bsuir.tattoo4u.service.UserService;
+import by.bsuir.tattoo4u.entity.*;
+import by.bsuir.tattoo4u.service.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -18,7 +13,9 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Controller
 @RequestMapping(value = "api")
@@ -26,12 +23,19 @@ public class PostController {
     private final PostService postService;
     private final TokenService tokenService;
     private final UserService userService;
+    private final PhotoService photoService;
 
     @Autowired
-    public PostController(PostService postService, TokenService tokenService, UserService userService) {
+    public PostController(
+            PostService postService,
+            TokenService tokenService,
+            UserService userService,
+            PhotoService photoService
+    ) {
         this.postService = postService;
         this.tokenService = tokenService;
         this.userService = userService;
+        this.photoService = photoService;
     }
 
     @PostMapping(value = "add-post", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
@@ -47,13 +51,14 @@ public class PostController {
 
         User user = userService.getByUsername(username);
 
-        Post post = new Post(postRequestDto.getDescription(), user, postRequestDto.getTags());
-
         PhotoUpload photoUpload = new PhotoUpload(postRequestDto.getFile());
 
         try {
-            postService.savePhoto(post, photoUpload);
+            Photo photo = photoService.save(photoUpload);
+
+            Post post = new Post(postRequestDto.getDescription(), user, photo, postRequestDto.getTags());
             postService.save(post);
+
             return new ResponseEntity<>(HttpStatus.CREATED);
         } catch (ServiceException e) {
             throw new ControllerException(e);
@@ -83,7 +88,7 @@ public class PostController {
 
                 return new ResponseEntity<>(postDtoList, HttpStatus.OK);
             } else {
-                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+                return new ResponseEntity<>("User with the specified id does not exist", HttpStatus.BAD_REQUEST);
             }
         } catch (ServiceException e) {
             throw new ControllerException(e);
@@ -104,13 +109,28 @@ public class PostController {
     }
 
     @DeleteMapping(value = "delete-post/{post}")
-    public ResponseEntity<?> deletePost(@PathVariable Post post) {
+    public ResponseEntity<?> deletePost(@PathVariable Post post, @RequestHeader("Authorization") String token) {
         try {
+            //token validation
+            token = token.substring(7); //move to service
+
+            String username = tokenService.getUsername(token);
+
+            User user = userService.getByUsername(username);
+
+            List<String> roles = user.getRoles().stream()
+                    .map(Role::getName)
+                    .collect(Collectors.toList());
+
             if (post != null) {
-                postService.delete(post);
-                return new ResponseEntity<>(HttpStatus.OK);
+                if (post.getAuthor().getId().equals(user.getId()) || roles.contains(RoleType.ADMIN.toString())) {
+                    postService.delete(post);
+                    return new ResponseEntity<>(HttpStatus.OK);
+                } else {
+                    return new ResponseEntity<>("Access is denied", HttpStatus.FORBIDDEN);
+                }
             } else {
-                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+                return new ResponseEntity<>("Post with the specified id does not exist", HttpStatus.BAD_REQUEST);
             }
         } catch (ServiceException e) {
             throw new ControllerException(e);
